# api.yaml
openapi: 3.0.0
info:
  title: Movix Auth Service API
  version: "1.0"
servers:
  - url: "https://api.example.com"
paths:
  /register:
    post:
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  required: true
                password:
                  type: string
                  required: true
                email:
                  type: string
                first_name:
                  type: string
                last_name:
                  type: string

      responses:
        "201":
          description: User registered successfully
        "400":
          description: Bad request (e.g., missing required fields)

  /login:
    post:
      summary: Login to the account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  required: true
                password:
                  type: string
                  required: true
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                  refresh_token:
                    type: string
        "401":
          description: Unauthorized (e.g., invalid credentials)

  /token/refresh:
    post:
      summary: Refresh access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                refresh_token:
                  type: string
      responses:
        "200":
          description: Access token refreshed
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string # проверяем подпись
                  refresh_token:
                    type: string # храним в бд id и в payload

        "401":
          description: Unauthorized (e.g., invalid refresh token)


  /logout:
    post:
      summary: Logout of the account
      responses:
        "204":
          description: Logout successful


  /user:
    get:
      summary: Get sensitive data of current user
      responses:
        "200":
          description: Data get successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  user_id:
                    type: string
                  username:
                    type: string
                  email:
                    type: string
                  first_name:
                    type: string
                  last_name:
                    type: string

        "403":
          description: Forbidden

    put:
      summary: Change login or password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
                password_old:
                  type: string
      responses:
        "200":
          description: Login or password changed successfully
        "400":
          description: Bad request (e.g., missing required fields)


  /user/sign-in-history:
    get:
      summary: Get user's account sign-in history
      responses:
        "200":
          description: Sign-in history retrieved
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    timestamp:
                      type: string
                      format: date-time
                    ip:
                      type: string


  /roles:
    post:
      summary: Create a role
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
      responses:
        "201":
          description: Role created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  name:
                    type: string

        "400":
          description: Bad request (e.g., missing required fields)


    get:
      summary: View all roles
      responses:
        "200":
          description: Roles retrieved
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                    name:
                      type: string

  /roles/{role_id}:
    parameters:
      - name: role_id
        in: path
        required: true
        schema:
          type: integer
    put:
      summary: Change a role
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type:
                    integer
                name:
                  type: string
      responses:
        "200":
          description: Role changed successfully
        "400":
          description: Bad request (e.g., missing required fields)

    delete:
      summary: Delete a role
      responses:
        "204":
          description: Role deleted successfully

  /users/{user_id}/roles/{role_id}:
    parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: integer
      - name: role_id
        in: path
        required: true
        schema:
          type: integer
    get:
      summary: Check if user is assigned to the role
      responses:
        "200":
          description: Role is assigned to the user
        "404":
          description: Role not found for the user
    post:
      summary: Assign a role to a user
      responses:
        "201":
          description: Role assigned successfully
        "400":
          description: Bad request (e.g., missing required fields)
    delete:
      summary: Unassign user's role
      responses:
        "204":
          description: Role unassigned successfully

  /users/{user_id}/roles:
    get:
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: integer
      summary: List the user's roles
      responses:
        "200":
          description: Roles retrieved
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                    name:
                      type: string
        "404":
          description: User not found

components:
  schemas: